+++ common/src/main/java/io/github/mortuusars/envelope/world/item/crafting/MailRecipe.java
@@ -3,11 +3,13 @@ package io.github.mortuusars.envelope.world.item.crafting;
 import io.github.mortuusars.envelope.Envelope;
 import io.github.mortuusars.envelope.world.item.component.PackageContents;
 import io.github.mortuusars.envelope.world.mail.address.type.EntityAddress;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.util.Mth;
-import net.minecraft.world.entity.ExperienceOrb;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.RecipeInput;
 import net.minecraft.world.item.crafting.RecipeType;
+import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
@@ -22,6 +24,15 @@ public interface MailRecipe extends Recipe<PackageRecipeInput> {
     @Override
     default @NotNull RecipeType<?> getType() {
         return Envelope.RecipeTypes.MAILING.get();
+    }
+
+    /**
+     * Extended version of {@link Recipe#assemble(RecipeInput, HolderLookup.Provider)} to allow custom logic dependent on a level.
+     * <br>
+     * This method should only be used one time, to get the actual result of the craft, as implementations can mutate the state.
+     */
+    default @NotNull ItemStack assembleFinal(PackageRecipeInput input, Level level) {
+        return assemble(input, level.registryAccess());
     }
 
     @Override