+++ common/src/main/java/io/github/mortuusars/envelope/world/mail/entity/mail_service/payback_department/PaybackDepartmentData.java
@@ -54,8 +54,6 @@ public class PaybackDepartmentData extends SavedData {
     }
 
     private static PaybackDepartmentData load(CompoundTag tag, HolderLookup.Provider registries) {
-
-
         return CODEC.decode(registries.createSerializationContext(NbtOps.INSTANCE), tag)
               .ifError(e -> Envelope.LOGGER.error("Cannot load PaybackDepartmentData: {}", e.message()))
               .result()